SentimentHub/SentimentHub.AI/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.21.0
webdriver-manager==4.0.1
uvicorn==0.29.0
fastapi==0.111.0
# transformers==4.41.1
# torch==2.3.0
openai==1.30.0
google-generativeai==0.5.4
python-dotenv==1.0.1
beautifulsoup4
